reorganized components file structure
