+++ src/components/Grids.tsx
@@ -9,12 +9,12 @@ import {
 } from "@material-ui/core";
 
 // Assets
-import { iTransaction } from "../../utils/types";
+import { iTransaction } from "../utils/types";
 import {
   CONFIG_MONTHS,
   CONFIG_GROUP_STRUCTURE,
   CONFIG_CATEGORY_TO_GROUP,
-} from "../../utils/configurations";
+} from "../utils/configurations";
 
 // Types
 interface iGridData {
